requirements.txt
flask==2.0.1
flask-socketio==5.1.1
opencv-python==4.9.0.80
python-dotenv==0.19.0
eventlet==0.33.0
Pillow==10.4.0 